Move provider enum into its own module
